% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMADADepartmentalForum.R
\docType{package}
\name{IMADADepartmentalForum}
\alias{IMADADepartmentalForum}
\title{IMADADepartmentalForum: A template pdf presentation for IMADA Departmental Fora}
\description{
This template creates an R-markdown file and a Latex preamble.
}
